requirements.txt
streamlit
pandas
gspread
google-auth
google-generativeai
plotly
python-dotenv
